web-panel/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator==2.1.0
Pillow==10.1.0
cryptography==41.0.8
pyotp==2.9.0
qrcode==7.4.2
schedule==1.2.0
requests==2.31.0
psutil==5.9.6
docker==6.1.3
paramiko==3.3.1
